requirements.txt
flask
python-dotenv
flask_mysqldb
psycopg2
flask-mysql
mysql-connector-python